% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.human.custom.R
\docType{data}
\name{md.human.custom}
\alias{md.human.custom}
\title{md.human.custom in a custom data frame (human)}
\format{An object of class \code{data.frame} with 612 rows and 3 columns.}
\usage{
md.human.custom
}
\arguments{
\item{cancer}{cancer-related gene signatures}

\item{cancer/PAM50}{50-gene classifier for four breast cancer types. Source,
zip file with data, analysis and results files: \url{https://genome.unc.edu/pubsup/breastGEO/PAM50.zip}}

\item{pharmacology}{drug metabolism-related gene signatures}

\item{pharmacology/ADME*}{core, extended and related drug-metabolizing genes.
Source: \url{http://pharmaadme.org/}}
}
\description{
md.human.custom in a custom data frame (human)
}
\keyword{datasets}

