% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.human}
\alias{msigdf.human}
\title{MSigDB in a data frame (human) by gene symbol}
\format{
A data frame with 4 variables: \code{category_code},
  \code{category_subcode}, \code{geneset}, and \code{symbol}.
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/}
}
\usage{
msigdf.human
}
\description{
A data frame containing the Molecular Signatures Database (MSigDB) gene sets for human, with gene symbols.
}
\examples{
head(msigdf.human)
}
\keyword{datasets}
