% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.mouse.urls}
\alias{msigdf.mouse.urls}
\title{Links to MSigDB website mouse gene sets}
\format{
A data frame with 4 variables: \code{category_code}, \code{category_subcode}, \code{geneset}, and \code{url}.
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/mouse/geneset/}
}
\usage{
msigdf.mouse.urls
}
\description{
Data frame linking each mouse gene set name to the MSigDB website.
}
\examples{
head(msigdf.mouse.urls)
}
\keyword{datasets}
