% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.urls}
\alias{msigdf.urls}
\title{Links to MSigDB website human gene sets}
\format{
A data frame with 4 variables: \code{category_code}, \code{category_subcode}, \code{geneset}, and \code{url}.
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/human/geneset/}
}
\usage{
msigdf.urls
}
\description{
Data frame linking each human gene set name to the MSigDB website.
}
\examples{
head(msigdf.urls)
}
\keyword{datasets}
