% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdf.human.R
\docType{data}
\name{msigdf.human}
\alias{msigdf.human}
\title{MSigDB in a data frame (human)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1881999 rows and 3 columns.}
\source{
\url{http://bioinf.wehi.edu.au/software/MSigDB/} and \url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.human
}
\description{
MSigDB in a data frame (human)
}
\keyword{datasets}

