% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.human}
\alias{msigdf.human}
\title{MSigDB in a data frame (human)}
\format{A data frame with 3 variables: \code{collection}, \code{geneset}, and \code{entrez}. See vignette for more details.}
\source{
\url{http://bioinf.wehi.edu.au/software/MSigDB/} and \url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.human
}
\description{
MSigDB in a data frame (human)
}
\keyword{datasets}
