% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.human}
\alias{msigdf.human}
\title{MSigDB in a data frame (human)}
\format{A data frame with 4 variables: \code{category_code}, \code{category_subcode}, \code{geneset} and \code{symbol}. See vignette for more details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.human
}
\description{
MSigDB in a data frame (human)
}
\keyword{datasets}
