% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.mouse}
\alias{msigdf.mouse}
\title{MSigDB in a data frame (mouse)}
\format{A data frame with 3 variables: \code{collection}, \code{geneset}, and \code{entrez}. See vignette for more details.}
\source{
\url{http://bioinf.wehi.edu.au/software/MSigDB/} and \url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.mouse
}
\description{
MSigDB in a data frame (mouse)
}
\keyword{datasets}
