% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdf.mouse.R
\docType{data}
\name{msigdf.mouse}
\alias{msigdf.mouse}
\title{MSigDB in a data frame (mouse)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3389316 rows and 3 columns.}
\source{
\url{http://bioinf.wehi.edu.au/software/MSigDB/} and \url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.mouse
}
\description{
MSigDB in a data frame (mouse)
}
\keyword{datasets}

