% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.mouse}
\alias{msigdf.mouse}
\title{MSigDB in a data frame (mouse) by gene symbol}
\format{
A data frame with 4 variables: \code{category_code}, \code{category_subcode}, \code{geneset}, and \code{symbol}.
}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.mouse
}
\description{
A data frame containing the Molecular Signatures Database (MSigDB) gene sets for mouse, with gene symbols.
}
\examples{
head(msigdf.mouse)
}
\keyword{datasets}
