% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.mouse}
\alias{msigdf.mouse}
\title{MSigDB in a data frame (mouse)}
\format{A data frame with 4 variables: \code{category_code}, \code{category_subcode}, \code{geneset} and \code{symbol}. See vignette for more details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.mouse
}
\description{
MSigDB in a data frame (mouse)
}
\keyword{datasets}
