% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdf.urls.R
\docType{data}
\name{msigdf.urls}
\alias{msigdf.urls}
\title{Links to MSigDB website}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13311 rows and 3 columns.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.urls
}
\description{
Data frame linking each gene set name to the MSigDB website
}
\keyword{datasets}

