% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdf.urls}
\alias{msigdf.urls}
\title{Links to MSigDB website}
\format{A data frame with 3 variables: \code{collection}, \code{geneset}, and \code{url}. See vignette for more details.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb/}
}
\usage{
msigdf.urls
}
\description{
Data frame linking each gene set name to the MSigDB website
}
\keyword{datasets}
